ref: Выполнил рефакторинг utils.js `Разбил файл utils.js на 3 модуля: common.js, points.js, render.js
