+++ src/view/trip-new-point.js
@@ -1,7 +1,7 @@
 import {nanoid} from "nanoid";
 import dayjs from "dayjs";
 import {DESTINATIONS_ARRAY} from "../const.js";
-import {getRandomInt} from "../utils.js";
+import {getRandomInt} from "./utils/common.js";
 import AbstractView from "../view/abstract.js";
 
 const generateDistDatalist = (pointsList) => {
